fix: refactor date handling for progress notes to prevent double timezone conversion
