+++ src/hooks/queries/useProgressNotes.ts
@@ -4,7 +4,7 @@ import { ProgressNote } from '@/types/project';
 import { pb } from '@/lib/pocketbase';
 import { queryKeys } from './queryKeys';
 import { requireValidAuthStore } from '@/utils/authGuards';
-import { toUserDateString } from '@/utils/timezoneUtils';
+import { formatDateForStorage } from '@/utils/dateFormatting';
 import { createLogger } from '@/utils/logger';
 
 const logger = createLogger('useProgressNotes');
@@ -88,53 +88,7 @@ export function useAddProgressNoteMutation(projectId: string) {
       // Check authentication
       requireValidAuthStore();
 
-      // Handle YYYY-MM-DD strings specially to prevent double timezone conversion
-      const formatProgressNoteDate = (
-        value: string | undefined,
-        userTimezone?: string
-      ): string | null => {
-        if (!value || value === '') {
-          logger.debug('📅 Progress note date formatting: null/empty value', {
-            value,
-            userTimezone,
-          });
-          return null;
-        }
-
-        logger.debug('📅 Progress note date formatting input', {
-          inputValue: value,
-          inputType: typeof value,
-          userTimezone,
-          isYYYYMMDD: /^\d{4}-\d{2}-\d{2}$/.test(value),
-        });
-
-        // For YYYY-MM-DD strings from HTML date inputs, treat as date-only values
-        // This prevents the double timezone conversion bug in toUserDateString()
-        if (typeof value === 'string' && /^\d{4}-\d{2}-\d{2}$/.test(value)) {
-          logger.debug('📅 Progress note date: using date-only format (no timezone conversion)', {
-            inputValue: value,
-            outputValue: value,
-            userTimezone,
-            reason: 'YYYY-MM-DD strings represent calendar dates, not moments in time',
-          });
-          return value; // Return as-is for date-only values
-        }
-
-        // For other date formats, use the timezone conversion utilities
-        const result = toUserDateString(value, userTimezone);
-
-        logger.debug('📅 Progress note date formatting during save', {
-          inputValue: value,
-          inputType: typeof value,
-          userTimezone,
-          outputValue: result,
-          isChanged: String(value) !== result,
-        });
-
-        return result;
-      };
-
-      const convertedDate = formatProgressNoteDate(noteData.date, noteData.userTimezone);
+      const convertedDate = formatDateForStorage(noteData.date, noteData.userTimezone);
 
       // Log what we're about to send to PocketBase
       logger.debug('📝 Progress note data being sent to PocketBase', {
